Barber (#448)
* Added @barbershop debug command and corresponding functionality

* Changed some CharGen types that were ushort to uints. (They were short for legacy reasons no longer needed)

* Cleaned up Handle function to be thread-safe

+++ Source/ACE.DatLoader/Entity/TextureMapChange.cs
@@ -10,7 +10,7 @@ namespace ACE.DatLoader.Entity
     public class TextureMapChange
     {
         public byte PartIndex { get; set; }
-        public ushort OldTexture { get; set; }
-        public ushort NewTexture { get; set; }
+        public uint OldTexture { get; set; }
+        public uint NewTexture { get; set; }
     }
 }
